+++ src/stack.cuh
@@ -3,12 +3,12 @@
 
 struct Stacks{
     volatile int * stacks;
-    volatile unsigned int * stacksNumDeletedVertices;
+    volatile uint64_t* stacksNumDeletedVertices;
     int minimum;
 };
 
 __device__ void popStack(unsigned int vertexNum, int* vertexDegrees_s, unsigned int* numDeletedVertices, volatile int * stackVertexDegrees, 
-    volatile unsigned int* stackNumDeletedVertices, int * stackTop){
+    volatile uint64_t* stackNumDeletedVertices, int * stackTop){
 
     for(unsigned int vertex = threadIdx.x; vertex < vertexNum; vertex += blockDim.x) {
         vertexDegrees_s[vertex] = stackVertexDegrees[(*stackTop)*vertexNum + vertex];
@@ -20,7 +20,7 @@ __device__ void popStack(unsigned int vertexNum, int* vertexDegrees_s, unsigned 
 }
 
 __device__ void pushStack(unsigned int vertexNum, int* vertexDegrees_s, unsigned int* numDeletedVertices, volatile int * stackVertexDegrees, 
-    volatile unsigned int* stackNumDeletedVertices, int * stackTop){
+    volatile uint64_t* stackNumDeletedVertices, int * stackTop){
 
     ++(*stackTop);
     for(unsigned int vertex = threadIdx.x; vertex < vertexNum ; vertex += blockDim.x) {
@@ -35,7 +35,7 @@ Stacks allocateStacks(int vertexNum, int numBlocks, unsigned int minimum){
     Stacks stacks;
 
     volatile int* stacks_d;
-    volatile unsigned int* stacksNumDeletedVertices_d;
+    volatile uint64_t* stacksNumDeletedVertices_d;
     cudaMalloc((void**) &stacks_d, (minimum + 1) * (vertexNum) * sizeof(int) * numBlocks);
     cudaMalloc((void**) &stacksNumDeletedVertices_d, (minimum + 1) * sizeof(unsigned int) * numBlocks);
